+++ lib/hooks/moduleloader/index.js
@@ -1,38 +1,9 @@
-/**
- * Return a list of the plugins of the proyect
- * @param  {Object} config sails object
- * @return {Array}   Array of plugin names
- */
-var pluginSystem = function(config){
-  var fs        = require('fs');
-  var file_path = config.appPath + '/config/plugins';
-  var _         = require('lodash');
-
-  // TODO: Regular expression?
-  if(fs.existsSync(file_path + '.js') || fs.existsSync(file_path + '.coffee')){
-
-    var list = _.map(require(file_path).plugins, function(plugin){
-      return  config.appPath + '/node_modules/' + plugin;
-    });
-
-    var result =  _.filter(list, function(plugin){
-      return fs.existsSync(plugin);
-    });
-
-    return result;
-  }
-
-  return [];
-};
-
-
 module.exports = function(sails) {
-
-
   /**
    * Module dependencies
    */
   var buildDictionary = require('sails-build-dictionary');
+  var pluginsHelper   = require('./plugins');
   var async           = require('async');
   var path            = require('path');
   var _               = require('lodash');
@@ -90,7 +61,7 @@ module.exports = function(sails) {
 
           // Server-Side Code
           //
-          plugins: pluginSystem(config),
+          plugins: pluginsHelper.initialize(config),
           // For `controllers` hook
           controllers: path.resolve(config.appPath, 'api/controllers'),
           // For `policies` hook
